+++ lib/mapbox_maps_flutter.dart
@@ -2,6 +2,7 @@ library mapbox_maps_flutter;
 
 import 'dart:async';
 import 'dart:convert';
+import 'dart:developer' as developer;
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';