+++ lib/src/mapbox_maps_options.dart
@@ -2,9 +2,10 @@ part of mapbox_maps_flutter;
 
 /// Class for Mapbox SDK settings management
 final class MapboxOptions {
-  static _MapboxOptions _options = _MapboxOptions();
-
-  MapboxOptions._() {}
+  static _MapboxOptions _options = () {
+    LogConfiguration._setupDebugLoggingIfNeeded();
+    return _MapboxOptions();
+  } ();
 
   /// The access token that is used to access resources provided by Mapbox services.
   static Future<String> getAccessToken() {